write Appraisable module -- tests passing for tshirt.rb
